Implement network-only process creation.
Added `netonly` flag to `make_token` command to specify whether we should spawn processes as netonly or normally.

+++ SaltedCaramel/Win32/advapi32.cs
@@ -115,7 +115,7 @@ namespace SaltedCaramel.Win32
             string lpUsername,
             string lpDomain,
             string lpPassword,
-            IntPtr dwLogonFlags,
+            uint dwLogonFlags,
             string lpApplicationName,
             string lpCommandLine,
             IntPtr dwCreationFlags,
